requirements.txt
Django==6.0
djangorestframework==3.20.0
pytz
sqlparse
